Split dbHelper and database lib
